umodeCrawler/requirements.txt
asgiref==3.6.0
black==23.3.0
bleach==6.0.0
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
Django==4.2.1
django-cors-headers==4.0.0
django-jazzmin==2.6.0
django-modeladmin-reorder==0.3.1
django-summernote==0.8.20.0
gunicorn==20.1.0
idna==3.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
urllib3==2.0.3
webencodings==0.5.1
whitenoise==6.4.0
